Re-organise Groovy lib/ classes
