+++ lib/NfcoreTemplate.groovy
@@ -1,8 +1,49 @@
 /*
- * Functions to be run on completion of pipeline
+ * This file holds several functions used within the nf-core pipeline template.
  */
 
-class Completion {
+import org.yaml.snakeyaml.Yaml
+
+class NfcoreTemplate {
+
+    /*
+     * Check AWS Batch related parameters have been specified correctly
+     */
+    public static void awsBatch(workflow, params) {
+        if (workflow.profile.contains('awsbatch')) {
+            assert (params.awsqueue && params.awsregion) : "Specify correct --awsqueue and --awsregion parameters on AWSBatch!"
+            // Check outdir paths to be S3 buckets if running on AWSBatch
+            // related: https://github.com/nextflow-io/nextflow/issues/813
+            assert params.outdir.startsWith('s3:')       : "Outdir not on S3 - specify S3 Bucket to run on AWSBatch!"
+            // Prevent trace files to be stored on S3 since S3 does not support rolling files.
+            assert !params.tracedir.startsWith('s3:')    :  "Specify a local tracedir or run without trace! S3 cannot be used for tracefiles."
+        }
+    }
+
+    /*
+     * Check params.hostnames
+     */
+    public static void hostName(workflow, params, log) {
+        Map colors = logColours(params.monochrome_logs)
+        if (params.hostnames) {
+            def hostname = "hostname".execute().text.trim()
+            params.hostnames.each { prof, hnames ->
+                hnames.each { hname ->
+                    if (hostname.contains(hname) && !workflow.profile.contains(prof)) {
+                        log.info "=${colors.yellow}====================================================${colors.reset}=\n" +
+                                  "${colors.yellow}WARN: You are running with `-profile $workflow.profile`\n" +
+                                  "      but your machine hostname is ${colors.white}'$hostname'${colors.reset}.\n" +
+                                  "      ${colors.yellow_bold}Please use `-profile $prof${colors.reset}`\n" +
+                                  "=${colors.yellow}====================================================${colors.reset}="
+                    }
+                }
+            }
+        }
+    }
+
+    /*
+     * Construct and send completion email
+     */
     public static void email(workflow, params, summary_params, projectDir, log, multiqc_report=[], fail_mapped_reads=[:]) {
 
         // Set up the e-mail variables
@@ -89,7 +130,7 @@ class Completion {
         def sendmail_html          = sendmail_template.toString()
 
         // Send the HTML e-mail
-        Map colors = Utils.logColours(params.monochrome_logs)
+        Map colors = logColours(params.monochrome_logs)
         if (email_address) {
             try {
                 if (params.plaintext_email) { throw GroovyException('Send plaintext e-mail, not HTML') }
@@ -118,8 +159,11 @@ class Completion {
         output_tf.withWriter { w -> w << email_txt }
     }
 
+    /*
+     * Print pipeline summary on completion
+     */
     public static void summary(workflow, params, log, fail_mapped_reads=[:], pass_mapped_reads=[:]) {
-        Map colors = Utils.logColours(params.monochrome_logs)
+        Map colors = logColours(params.monochrome_logs)
 
         if (pass_mapped_reads.size() > 0) {
             def idx = 0
@@ -150,8 +194,103 @@ class Completion {
                 log.info "-${colors.purple}[$workflow.manifest.name]${colors.red} Pipeline completed successfully, but with errored process(es) ${colors.reset}-"
             }
         } else {
-            Checks.hostName(workflow, params, log)
+            hostName(workflow, params, log)
             log.info "-${colors.purple}[$workflow.manifest.name]${colors.red} Pipeline completed with errors${colors.reset}-"
         }
+    }
+
+    /*
+     * ANSII Colours used for terminal logging
+     */
+    public static Map logColours(Boolean monochrome_logs) {
+        Map colorcodes = [:]
+
+        // Reset / Meta
+        colorcodes['reset']      = monochrome_logs ? '' : "\033[0m"
+        colorcodes['bold']       = monochrome_logs ? '' : "\033[1m"
+        colorcodes['dim']        = monochrome_logs ? '' : "\033[2m"
+        colorcodes['underlined'] = monochrome_logs ? '' : "\033[4m"
+        colorcodes['blink']      = monochrome_logs ? '' : "\033[5m"
+        colorcodes['reverse']    = monochrome_logs ? '' : "\033[7m"
+        colorcodes['hidden']     = monochrome_logs ? '' : "\033[8m"
+
+        // Regular Colors
+        colorcodes['black']      = monochrome_logs ? '' : "\033[0;30m"
+        colorcodes['red']        = monochrome_logs ? '' : "\033[0;31m"
+        colorcodes['green']      = monochrome_logs ? '' : "\033[0;32m"
+        colorcodes['yellow']     = monochrome_logs ? '' : "\033[0;33m"
+        colorcodes['blue']       = monochrome_logs ? '' : "\033[0;34m"
+        colorcodes['purple']     = monochrome_logs ? '' : "\033[0;35m"
+        colorcodes['cyan']       = monochrome_logs ? '' : "\033[0;36m"
+        colorcodes['white']      = monochrome_logs ? '' : "\033[0;37m"
+
+        // Bold
+        colorcodes['bblack']     = monochrome_logs ? '' : "\033[1;30m"
+        colorcodes['bred']       = monochrome_logs ? '' : "\033[1;31m"
+        colorcodes['bgreen']     = monochrome_logs ? '' : "\033[1;32m"
+        colorcodes['byellow']    = monochrome_logs ? '' : "\033[1;33m"
+        colorcodes['bblue']      = monochrome_logs ? '' : "\033[1;34m"
+        colorcodes['bpurple']    = monochrome_logs ? '' : "\033[1;35m"
+        colorcodes['bcyan']      = monochrome_logs ? '' : "\033[1;36m"
+        colorcodes['bwhite']     = monochrome_logs ? '' : "\033[1;37m"
+
+        // Underline
+        colorcodes['ublack']     = monochrome_logs ? '' : "\033[4;30m"
+        colorcodes['ured']       = monochrome_logs ? '' : "\033[4;31m"
+        colorcodes['ugreen']     = monochrome_logs ? '' : "\033[4;32m"
+        colorcodes['uyellow']    = monochrome_logs ? '' : "\033[4;33m"
+        colorcodes['ublue']      = monochrome_logs ? '' : "\033[4;34m"
+        colorcodes['upurple']    = monochrome_logs ? '' : "\033[4;35m"
+        colorcodes['ucyan']      = monochrome_logs ? '' : "\033[4;36m"
+        colorcodes['uwhite']     = monochrome_logs ? '' : "\033[4;37m"
+
+        // High Intensity
+        colorcodes['iblack']     = monochrome_logs ? '' : "\033[0;90m"
+        colorcodes['ired']       = monochrome_logs ? '' : "\033[0;91m"
+        colorcodes['igreen']     = monochrome_logs ? '' : "\033[0;92m"
+        colorcodes['iyellow']    = monochrome_logs ? '' : "\033[0;93m"
+        colorcodes['iblue']      = monochrome_logs ? '' : "\033[0;94m"
+        colorcodes['ipurple']    = monochrome_logs ? '' : "\033[0;95m"
+        colorcodes['icyan']      = monochrome_logs ? '' : "\033[0;96m"
+        colorcodes['iwhite']     = monochrome_logs ? '' : "\033[0;97m"
+
+        // Bold High Intensity
+        colorcodes['biblack']    = monochrome_logs ? '' : "\033[1;90m"
+        colorcodes['bired']      = monochrome_logs ? '' : "\033[1;91m"
+        colorcodes['bigreen']    = monochrome_logs ? '' : "\033[1;92m"
+        colorcodes['biyellow']   = monochrome_logs ? '' : "\033[1;93m"
+        colorcodes['biblue']     = monochrome_logs ? '' : "\033[1;94m"
+        colorcodes['bipurple']   = monochrome_logs ? '' : "\033[1;95m"
+        colorcodes['bicyan']     = monochrome_logs ? '' : "\033[1;96m"
+        colorcodes['biwhite']    = monochrome_logs ? '' : "\033[1;97m"
+
+        return colorcodes
+    }
+
+    /*
+     * Does what is says on the tin
+     */
+    public static String dashedLine(monochrome_logs) {
+        Map colors = logColours(monochrome_logs)
+        return "-${colors.dim}----------------------------------------------------${colors.reset}-"
+    }
+
+    /*
+     * nf-core logo
+     */         
+    public static String logo(workflow, monochrome_logs) {
+        Map colors = logColours(monochrome_logs)
+        String.format(
+            """\n
+            ${dashedLine(monochrome_logs)}
+                                                    ${colors.green},--.${colors.black}/${colors.green},-.${colors.reset}
+            ${colors.blue}        ___     __   __   __   ___     ${colors.green}/,-._.--~\'${colors.reset}
+            ${colors.blue}  |\\ | |__  __ /  ` /  \\ |__) |__         ${colors.yellow}}  {${colors.reset}
+            ${colors.blue}  | \\| |       \\__, \\__/ |  \\ |___     ${colors.green}\\`-._,-`-,${colors.reset}
+                                                    ${colors.green}`._,._,\'${colors.reset}
+            ${colors.purple}  ${workflow.manifest.name} v${workflow.manifest.version}${colors.reset}
+            ${dashedLine(monochrome_logs)}
+            """.stripIndent()
+        )
     }
 }
