+++ lib/Utils.groovy
@@ -1,94 +1,46 @@
 /*
- * This file holds several Groovy functions used across the nf-core pipeline template
+ * This file holds several Groovy functions that could be useful for any Nextflow pipeline
  */
 
+import org.yaml.snakeyaml.Yaml
+
 class Utils {
-    public static Map logColours(Boolean monochrome_logs) {
-        Map colorcodes = [:]
 
-        // Reset / Meta
-        colorcodes['reset']      = monochrome_logs ? '' : "\033[0m"
-        colorcodes['bold']       = monochrome_logs ? '' : "\033[1m"
-        colorcodes['dim']        = monochrome_logs ? '' : "\033[2m"
-        colorcodes['underlined'] = monochrome_logs ? '' : "\033[4m"
-        colorcodes['blink']      = monochrome_logs ? '' : "\033[5m"
-        colorcodes['reverse']    = monochrome_logs ? '' : "\033[7m"
-        colorcodes['hidden']     = monochrome_logs ? '' : "\033[8m"
+    /*
+     * When running with -profile conda, warn if channels have not been set-up appropriately
+     */
+    public static void checkCondaChannels(log) {
+        Yaml parser = new Yaml()
+        def channels = []
+        try {
+            def config = parser.load("conda config --show channels".execute().text)
+            channels = config.channels
+        } catch(NullPointerException | IOException e) {
+            log.warn "Could not verify conda channel configuration."
+            return
+        }
 
-        // Regular Colors
-        colorcodes['black']      = monochrome_logs ? '' : "\033[0;30m"
-        colorcodes['red']        = monochrome_logs ? '' : "\033[0;31m"
-        colorcodes['green']      = monochrome_logs ? '' : "\033[0;32m"
-        colorcodes['yellow']     = monochrome_logs ? '' : "\033[0;33m"
-        colorcodes['blue']       = monochrome_logs ? '' : "\033[0;34m"
-        colorcodes['purple']     = monochrome_logs ? '' : "\033[0;35m"
-        colorcodes['cyan']       = monochrome_logs ? '' : "\033[0;36m"
-        colorcodes['white']      = monochrome_logs ? '' : "\033[0;37m"
+        // Check that all channels are present
+        def required_channels = ['conda-forge', 'bioconda', 'defaults']
+        def conda_check_failed = !required_channels.every { ch -> ch in channels }
 
-        // Bold
-        colorcodes['bblack']     = monochrome_logs ? '' : "\033[1;30m"
-        colorcodes['bred']       = monochrome_logs ? '' : "\033[1;31m"
-        colorcodes['bgreen']     = monochrome_logs ? '' : "\033[1;32m"
-        colorcodes['byellow']    = monochrome_logs ? '' : "\033[1;33m"
-        colorcodes['bblue']      = monochrome_logs ? '' : "\033[1;34m"
-        colorcodes['bpurple']    = monochrome_logs ? '' : "\033[1;35m"
-        colorcodes['bcyan']      = monochrome_logs ? '' : "\033[1;36m"
-        colorcodes['bwhite']     = monochrome_logs ? '' : "\033[1;37m"
+        // Check that they are in the right order
+        conda_check_failed |= !(channels.indexOf('conda-forge') < channels.indexOf('bioconda'))
+        conda_check_failed |= !(channels.indexOf('bioconda') < channels.indexOf('defaults'))
 
-        // Underline
-        colorcodes['ublack']     = monochrome_logs ? '' : "\033[4;30m"
-        colorcodes['ured']       = monochrome_logs ? '' : "\033[4;31m"
-        colorcodes['ugreen']     = monochrome_logs ? '' : "\033[4;32m"
-        colorcodes['uyellow']    = monochrome_logs ? '' : "\033[4;33m"
-        colorcodes['ublue']      = monochrome_logs ? '' : "\033[4;34m"
-        colorcodes['upurple']    = monochrome_logs ? '' : "\033[4;35m"
-        colorcodes['ucyan']      = monochrome_logs ? '' : "\033[4;36m"
-        colorcodes['uwhite']     = monochrome_logs ? '' : "\033[4;37m"
-
-        // High Intensity
-        colorcodes['iblack']     = monochrome_logs ? '' : "\033[0;90m"
-        colorcodes['ired']       = monochrome_logs ? '' : "\033[0;91m"
-        colorcodes['igreen']     = monochrome_logs ? '' : "\033[0;92m"
-        colorcodes['iyellow']    = monochrome_logs ? '' : "\033[0;93m"
-        colorcodes['iblue']      = monochrome_logs ? '' : "\033[0;94m"
-        colorcodes['ipurple']    = monochrome_logs ? '' : "\033[0;95m"
-        colorcodes['icyan']      = monochrome_logs ? '' : "\033[0;96m"
-        colorcodes['iwhite']     = monochrome_logs ? '' : "\033[0;97m"
-
-        // Bold High Intensity
-        colorcodes['biblack']    = monochrome_logs ? '' : "\033[1;90m"
-        colorcodes['bired']      = monochrome_logs ? '' : "\033[1;91m"
-        colorcodes['bigreen']    = monochrome_logs ? '' : "\033[1;92m"
-        colorcodes['biyellow']   = monochrome_logs ? '' : "\033[1;93m"
-        colorcodes['biblue']     = monochrome_logs ? '' : "\033[1;94m"
-        colorcodes['bipurple']   = monochrome_logs ? '' : "\033[1;95m"
-        colorcodes['bicyan']     = monochrome_logs ? '' : "\033[1;96m"
-        colorcodes['biwhite']    = monochrome_logs ? '' : "\033[1;97m"
-
-        return colorcodes
+        if (conda_check_failed) {
+            log.warn "=============================================================================\n" +
+                     "  There is a problem with your Conda configuration!\n\n" + 
+                     "  You will need to set-up the conda-forge and bioconda channels correctly.\n" +
+                     "  Please refer to https://bioconda.github.io/user/install.html#set-up-channels\n" +
+                     "  NB: The order of the channels matters!\n" +
+                     "==================================================================================="
+        }
     }
 
-    public static String dashedLine(monochrome_logs) {
-        Map colors = logColours(monochrome_logs)
-        return "-${colors.dim}----------------------------------------------------${colors.reset}-"
-    }
-
-    public static String logo(workflow, monochrome_logs) {
-        Map colors = logColours(monochrome_logs)
-        String.format(
-            """\n
-            ${dashedLine(monochrome_logs)}
-                                                    ${colors.green},--.${colors.black}/${colors.green},-.${colors.reset}
-            ${colors.blue}        ___     __   __   __   ___     ${colors.green}/,-._.--~\'${colors.reset}
-            ${colors.blue}  |\\ | |__  __ /  ` /  \\ |__) |__         ${colors.yellow}}  {${colors.reset}
-            ${colors.blue}  | \\| |       \\__, \\__/ |  \\ |___     ${colors.green}\\`-._,-`-,${colors.reset}
-                                                    ${colors.green}`._,._,\'${colors.reset}
-            ${colors.purple}  ${workflow.manifest.name} v${workflow.manifest.version}${colors.reset}
-            ${dashedLine(monochrome_logs)}
-            """.stripIndent()
-        )
-    }
-
+    /*
+     * Join module args with appropriate spacing
+     */
     public static String joinModuleArgs(args_list) {
         return ' ' + args_list.join(' ')
     }
